Verdi 2.2 changes

+++ verdi_core/src/anl/verdi/core/VerdiConstants.java
@@ -75,9 +75,9 @@ public interface VerdiConstants {
 	String TIME_STEP = "ts".toUpperCase();
 	String LAYER = "layer".toUpperCase();
 
-	public static final float NC_FILL_FLOAT = 9.9692099683868690e+36f;		// 2014 changed from -9.999E36f to assist floating point comparison
-	public static final float BADVAL3 = -9.998E36f;		// 2014 changed from -9.999E36f to assist floating point comparison
-	public static final float AMISS3  = -8.999E36f;		// 2014 changed from -9.000E36f to assist floating point comparison
+	public static final double NC_FILL_FLOAT = 9.9692099683868690e+36;		// 2014 changed from -9.999E36f to assist floating point comparison
+	public static final double BADVAL3 = -9.998E36;		// 2014 changed from -9.999E36f to assist floating point comparison
+	public static final double AMISS3  = -8.999E36;		// 2014 changed from -9.000E36f to assist floating point comparison
 	
 	public static final int NETCDF_CONV_ARW_WRF = 10;
